requirements.txt
PyQt6
opencv-python
numpy
pandas
openpyxl
requests
Pillow
pyinstaller